move gui helper class to better location
