+++ src/main/java/de/hysky/skyblocker/utils/render/HudHelper.java
@@ -1,9 +1,11 @@
 package de.hysky.skyblocker.utils.render;
 
 import java.awt.Color;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import de.hysky.skyblocker.utils.render.gui.state.CustomShapeGuiElementRenderState;
 import org.joml.Matrix3x2f;
 
 import de.hysky.skyblocker.utils.render.gui.state.EquipmentGuiElementRenderState;
@@ -25,20 +27,31 @@ import net.minecraft.registry.RegistryKey;
 import net.minecraft.text.OrderedText;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.ColorHelper;
+import org.joml.Vector2f;
 
 public class HudHelper {
 	private static final MinecraftClient CLIENT = MinecraftClient.getInstance();
 
-    public static void renderNineSliceColored(DrawContext context, Identifier texture, int x, int y, int width, int height, int argb) {
-        context.drawGuiTexture(RenderPipelines.GUI_TEXTURED, texture, x, y, width, height, argb);
-    }
+	public static void renderNineSliceColored(DrawContext context, Identifier texture, int x, int y, int width, int height, int argb) {
+		context.drawGuiTexture(RenderPipelines.GUI_TEXTURED, texture, x, y, width, height, argb);
+	}
 
-    public static void renderNineSliceColored(DrawContext context, Identifier texture, int x, int y, int width, int height, Color color) {
-        renderNineSliceColored(context, texture, x, y, width, height, ColorHelper.getArgb(color.getAlpha(), color.getRed(), color.getGreen(), color.getBlue()));
-    }
+	public static void renderNineSliceColored(DrawContext context, Identifier texture, int x, int y, int width, int height, Color color) {
+		renderNineSliceColored(context, texture, x, y, width, height, ColorHelper.getArgb(color.getAlpha(), color.getRed(), color.getGreen(), color.getBlue()));
+	}
 
 	public static void drawHorizontalGradient(DrawContext context, float startX, float startY, float endX, float endY, int colorStart, int colorEnd) {
 		context.state.addSimpleElement(new HorizontalGradientGuiElementRenderState(RenderPipelines.GUI, TextureSetup.empty(), new Matrix3x2f(context.getMatrices()), (int) startX, (int) startY, (int) endX, (int) endY, colorStart, colorEnd, context.scissorStack.peekLast()));
+	}
+
+	/**
+	 * Draws shape with given vertices. Note the vertices must be in the right or for this to work properly this function does not sort them for you.
+	 * @param context draw context
+	 * @param vertices vertices of shape
+	 * @param color color of shape
+	 */
+	public static void drawCustomShape(DrawContext context, List<Vector2f> vertices, int color) {
+		context.state.addSimpleElement(new CustomShapeGuiElementRenderState(RenderPipelines.GUI, TextureSetup.empty(), new Matrix3x2f(context.getMatrices()), vertices, color, context.scissorStack.peekLast()));
 	}
 
 	/**
@@ -61,7 +74,7 @@ public class HudHelper {
 		context.state.addText(renderState);
 	}
 
-    public static boolean pointIsInArea(double x, double y, double x1, double y1, double x2, double y2) {
-        return x >= x1 && x <= x2 && y >= y1 && y <= y2;
-    }
+	public static boolean pointIsInArea(double x, double y, double x1, double y1, double x2, double y2) {
+		return x >= x1 && x <= x2 && y >= y1 && y <= y2;
+	}
 }
